+++ src/batch_hasher.rs
@@ -6,25 +6,16 @@ use crate::error::{ClError, Error};
 use crate::poseidon::SimplePoseidonBatchHasher;
 #[cfg(any(feature = "cuda", feature = "opencl"))]
 use crate::proteus::gpu::ClBatchHasher;
-#[cfg(feature = "futhark")]
-use crate::triton::{cl, gpu::GpuBatchHasher};
 use crate::{Arity, BatchHasher, Strength, DEFAULT_STRENGTH};
 use blstrs::Scalar as Fr;
 use generic_array::GenericArray;
-#[cfg(feature = "futhark")]
-use rust_gpu_tools::opencl;
 use rust_gpu_tools::Device;
-
-#[cfg(feature = "futhark")]
-use triton::FutharkContext;
 
 pub enum Batcher<A>
 where
     A: Arity<Fr>,
 {
     Cpu(SimplePoseidonBatchHasher<A>),
-    #[cfg(feature = "futhark")]
-    OpenCl(GpuBatchHasher<A>),
     #[cfg(any(feature = "cuda", feature = "opencl"))]
     OpenCl(ClBatchHasher<A>),
 }
@@ -47,17 +38,6 @@ where
     }
 
     /// Create a new GPU batcher for an arbitrarily picked device.
-    #[cfg(feature = "futhark")]
-    pub fn pick_gpu(max_batch_size: usize) -> Result<Self, Error> {
-        let futhark_context = cl::default_futhark_context()?;
-        Ok(Self::OpenCl(GpuBatchHasher::<A>::new_with_strength(
-            futhark_context,
-            DEFAULT_STRENGTH,
-            max_batch_size,
-        )?))
-    }
-
-    /// Create a new GPU batcher for an arbitrarily picked device.
     #[cfg(any(feature = "cuda", feature = "opencl"))]
     pub fn pick_gpu(max_batch_size: usize) -> Result<Self, Error> {
         let device = *Device::all()
@@ -66,34 +46,10 @@ where
         Self::new(device, max_batch_size)
     }
 
-    #[cfg(feature = "futhark")]
-    /// Create a new GPU batcher for a certain device.
-    pub fn new(device: &Device, max_batch_size: usize) -> Result<Self, Error> {
-        Self::with_strength(device, DEFAULT_STRENGTH, max_batch_size)
-    }
-
     #[cfg(any(feature = "cuda", feature = "opencl"))]
     /// Create a new GPU batcher for a certain device.
     pub fn new(device: &Device, max_batch_size: usize) -> Result<Self, Error> {
         Self::with_strength(device, DEFAULT_STRENGTH, max_batch_size)
-    }
-
-    #[cfg(feature = "futhark")]
-    /// Create a new GPU batcher for a certain device with a specified strength.
-    pub fn with_strength(
-        device: &Device,
-        strength: Strength,
-        max_batch_size: usize,
-    ) -> Result<Self, Error> {
-        let opencl_device = device
-            .opencl_device()
-            .ok_or(Error::ClError(ClError::DeviceNotFound))?;
-        let futhark_context = cl::futhark_context(&opencl_device)?;
-        Ok(Self::OpenCl(GpuBatchHasher::<A>::new_with_strength(
-            futhark_context,
-            strength,
-            max_batch_size,
-        )?))
     }
 
     #[cfg(any(feature = "cuda", feature = "opencl"))]
@@ -118,7 +74,7 @@ where
     fn hash(&mut self, preimages: &[GenericArray<Fr, A>]) -> Result<Vec<Fr>, Error> {
         match self {
             Batcher::Cpu(batcher) => batcher.hash(preimages),
-            #[cfg(any(feature = "futhark", feature = "cuda", feature = "opencl"))]
+            #[cfg(any(feature = "cuda", feature = "opencl"))]
             Batcher::OpenCl(batcher) => batcher.hash(preimages),
         }
     }
@@ -126,7 +82,7 @@ where
     fn max_batch_size(&self) -> usize {
         match self {
             Batcher::Cpu(batcher) => batcher.max_batch_size(),
-            #[cfg(any(feature = "futhark", feature = "cuda", feature = "opencl"))]
+            #[cfg(any(feature = "cuda", feature = "opencl"))]
             Batcher::OpenCl(batcher) => batcher.max_batch_size(),
         }
     }