+++ src/column_tree_builder.rs
@@ -56,7 +56,7 @@ where
             }
             None => columns.iter().enumerate().for_each(|(i, column)| {
                 self.data[start + i] =
-                    Poseidon::new_with_preimage(&column, &self.column_constants).hash();
+                    Poseidon::new_with_preimage(column, &self.column_constants).hash();
             }),
         };
 
@@ -141,10 +141,7 @@ where
     }
 }
 
-#[cfg(all(
-    any(feature = "futhark", feature = "cuda", feature = "opencl"),
-    not(target_os = "macos")
-))]
+#[cfg(all(any(feature = "cuda", feature = "opencl"), not(target_os = "macos")))]
 #[cfg(test)]
 mod tests {
     use super::*;