+++ src/error.rs
@@ -1,9 +1,7 @@
-#[cfg(feature = "futhark")]
-use crate::triton::cl;
 use std::{error, fmt};
 
 #[derive(Debug, Clone)]
-#[cfg(any(feature = "futhark", feature = "cuda", feature = "opencl"))]
+#[cfg(any(feature = "cuda", feature = "opencl"))]
 pub enum ClError {
     DeviceNotFound,
     PlatformNotFound,
@@ -16,10 +14,10 @@ pub enum ClError {
     GetDeviceError,
 }
 
-#[cfg(any(feature = "futhark", feature = "cuda", feature = "opencl"))]
+#[cfg(any(feature = "cuda", feature = "opencl"))]
 pub type ClResult<T> = std::result::Result<T, ClError>;
 
-#[cfg(any(feature = "futhark", feature = "cuda", feature = "opencl"))]
+#[cfg(any(feature = "cuda", feature = "opencl"))]
 impl fmt::Display for ClError {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         match self {
@@ -50,26 +48,9 @@ pub enum Error {
     /// Attempt to reference an index element that is out of bounds
     IndexOutOfBounds,
     GpuError(String),
-    #[cfg(any(feature = "futhark", feature = "cuda", feature = "opencl"))]
+    #[cfg(any(feature = "cuda", feature = "opencl"))]
     ClError(ClError),
-    #[cfg(feature = "futhark")]
-    TritonError(String),
-    DecodingError,
     Other(String),
-}
-
-#[cfg(feature = "futhark")]
-impl From<ClError> for Error {
-    fn from(e: ClError) -> Self {
-        Self::ClError(e)
-    }
-}
-
-#[cfg(feature = "futhark")]
-impl From<triton::Error> for Error {
-    fn from(e: triton::Error) -> Self {
-        Self::TritonError(e.to_string())
-    }
 }
 
 #[cfg(any(feature = "cuda", feature = "opencl"))]
@@ -90,11 +71,8 @@ impl fmt::Display for Error {
             ),
             Error::IndexOutOfBounds => write!(f, "The referenced index is outs of bounds."),
             Error::GpuError(s) => write!(f, "GPU Error: {}", s),
-            #[cfg(any(feature = "futhark", feature = "cuda", feature = "opencl"))]
+            #[cfg(any(feature = "cuda", feature = "opencl"))]
             Error::ClError(e) => write!(f, "OpenCL Error: {}", e),
-            #[cfg(feature = "futhark")]
-            Error::TritonError(e) => write!(f, "Neptune-triton Error: {}", e),
-            Error::DecodingError => write!(f, "PrimeFieldDecodingError"),
             Error::Other(s) => write!(f, "{}", s),
         }
     }